refactor(checker): split JSX union prop helpers
